Add UT for ORTModule init fallback
