+++ orttraining/orttraining/python/training/ortmodule/__init__.py
@@ -6,10 +6,10 @@
 import os
 import sys
 
-from glob import glob
 from packaging import version
 
-from ._fallback import _FallbackManager, ORTModuleInitException, wrap_exception
+from ._fallback import _FallbackManager, ORTModuleFallbackException, ORTModuleInitException, wrap_exception
+from .torch_cpp_extensions import is_installed as is_torch_cpp_extensions_installed
 
 
 ################################################################################
@@ -19,6 +19,7 @@ ONNX_OPSET_VERSION = 12
 MINIMUM_RUNTIME_PYTORCH_VERSION_STR = '1.8.1'
 TORCH_CPP_DIR = os.path.join(os.path.dirname(__file__),
                              'torch_cpp_extensions')
+FALLBACK_INIT_EXCEPTION = None
 
 # Verify minimum PyTorch version is installed before proceding to ONNX Runtime initialization
 try:
@@ -30,18 +31,19 @@ try:
                              RuntimeError(
                                  f'ONNX Runtime ORTModule frontend requires PyTorch version greater or equal to {MINIMUM_RUNTIME_PYTORCH_VERSION_STR}, '
                                  f'but version {torch.__version__} was found instead.'))
-except:
-    raise wrap_exception(ORTModuleInitException,
-                         RuntimeError(f'PyTorch {MINIMUM_RUNTIME_PYTORCH_VERSION_STR} must be installed in order to run ONNX Runtime ORTModule frontend!'))
+except ORTModuleFallbackException as e:
+    # Initialization fallback is handled at ORTModule.__init__
+    FALLBACK_INIT_EXCEPTION = e
+except ImportError as e:
+    raise RuntimeError(f'PyTorch {MINIMUM_RUNTIME_PYTORCH_VERSION_STR} must be installed in order to run ONNX Runtime ORTModule frontend!') from e
 
 # Verify whether PyTorch C++ extensions are already compiled
-torch_cpp_exts = glob(os.path.join(TORCH_CPP_DIR, '*.so'))
-torch_cpp_exts.extend(glob(os.path.join(TORCH_CPP_DIR, '*.dll')))
-torch_cpp_exts.extend(glob(os.path.join(TORCH_CPP_DIR, '*.dylib')))
-if not torch_cpp_exts and '-m' not in sys.argv:
-    raise wrap_exception(ORTModuleInitException,
-                         EnvironmentError(f"ORTModule's extensions were not detected at '{TORCH_CPP_DIR}' folder. "
-                                         "Run `python -m torch_ort.configure` before using `ORTModule` frontend."))
+
+if not is_torch_cpp_extensions_installed(TORCH_CPP_DIR) and '-m' not in sys.argv:
+    FALLBACK_INIT_EXCEPTION = wrap_exception(ORTModuleInitException,
+                                                        EnvironmentError(
+                                                            f"ORTModule's extensions were not detected at '{TORCH_CPP_DIR}' folder. "
+                                                            "Run `python -m torch_ort.configure` before using `ORTModule` frontend."))
 
 # PyTorch custom Autograd function support
 from ._custom_autograd_function import enable_custom_autograd_support
